requirements.txt
PyVirtualDisplay
selenium
obs-websocket-py
psutil
pulsectl
retrying
